requirements.txt
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
